+++ weaviate/collection/grpc/__init__.py
@@ -1,0 +1,8 @@
+from .batch import _BatchGRPC
+from .query import _GrpcCollection, _GrpcCollectionModel
+
+__all__ = [
+    "_BatchGRPC",
+    "_GrpcCollection",
+    "_GrpcCollectionModel",
+]
