+++ weaviate/collection/grpc/base/shared.py
