+++ weaviate/collection/grpc/queries/near_audio.py
@@ -1,0 +1,44 @@
+from typing import (
+    Optional,
+    Union,
+    Type,
+)
+
+from weaviate.collection.classes.filters import (
+    _Filters,
+)
+from weaviate.collection.classes.grpc import (
+    MetadataQuery,
+    PROPERTIES,
+)
+from weaviate.collection.classes.internal import (
+    _QueryReturn,
+)
+from weaviate.collection.classes.types import (
+    Properties,
+)
+from weaviate.collection.grpc.base.wrapper import _Grpc
+
+
+class _NearAudio(_Grpc):
+    def near_audio(
+        self,
+        near_audio: str,
+        certainty: Optional[float] = None,
+        distance: Optional[float] = None,
+        auto_limit: Optional[int] = None,
+        filters: Optional[_Filters] = None,
+        return_metadata: Optional[MetadataQuery] = None,
+        return_properties: Optional[Union[PROPERTIES, Type[Properties]]] = None,
+    ) -> _QueryReturn[Properties]:
+        ret_properties, ret_type = self._determine_generic(return_properties)
+        res = self._query().near_audio(
+            audio=near_audio,
+            certainty=certainty,
+            distance=distance,
+            filters=filters,
+            autocut=auto_limit,
+            return_metadata=return_metadata,
+            return_properties=ret_properties,
+        )
+        return self._result_to_query_return(res, ret_type)
