+++ weaviate/collection/grpc/queries/near_object.py
@@ -1,0 +1,48 @@
+from typing import (
+    Optional,
+    Union,
+    Type,
+)
+
+from weaviate.collection.classes.filters import (
+    _Filters,
+)
+from weaviate.collection.classes.grpc import (
+    GroupBy,
+    MetadataQuery,
+    PROPERTIES,
+)
+from weaviate.collection.classes.internal import (
+    _QueryReturn,
+)
+from weaviate.collection.classes.types import (
+    Properties,
+)
+from weaviate.collection.grpc.base.wrapper import _Grpc
+from weaviate.weaviate_types import UUID
+
+
+class _NearObject(_Grpc):
+    def near_object(
+        self,
+        near_object: UUID,
+        certainty: Optional[float] = None,
+        distance: Optional[float] = None,
+        auto_limit: Optional[int] = None,
+        filters: Optional[_Filters] = None,
+        group_by: Optional[GroupBy] = None,
+        return_metadata: Optional[MetadataQuery] = None,
+        return_properties: Optional[Union[PROPERTIES, Type[Properties]]] = None,
+    ) -> _QueryReturn[Properties]:
+        ret_properties, ret_type = self._determine_generic(return_properties)
+        res = self._query().near_object(
+            near_object=near_object,
+            certainty=certainty,
+            distance=distance,
+            autocut=auto_limit,
+            filters=filters,
+            group_by=group_by,
+            return_metadata=return_metadata,
+            return_properties=ret_properties,
+        )
+        return self._result_to_query_return(res, ret_type)
