+++ weaviate/collection/grpc/queries/near_text.py
@@ -1,0 +1,121 @@
+from typing import (
+    List,
+    Literal,
+    Optional,
+    Union,
+    Type,
+    overload,
+)
+
+from weaviate.collection.classes.filters import (
+    _Filters,
+)
+from weaviate.collection.classes.grpc import (
+    Generate,
+    GroupBy,
+    MetadataQuery,
+    PROPERTIES,
+    Move,
+)
+from weaviate.collection.classes.internal import (
+    _GenerativeReturn,
+    _GroupByReturn,
+    _QueryReturn,
+)
+from weaviate.collection.classes.types import (
+    Properties,
+)
+from weaviate.collection.grpc.base.wrapper import _Grpc
+
+
+class _NearText(_Grpc):
+    @overload
+    def near_text(
+        self,
+        query: Union[List[str], str],
+        certainty: Optional[float] = None,
+        distance: Optional[float] = None,
+        move_to: Optional[Move] = None,
+        move_away: Optional[Move] = None,
+        auto_limit: Optional[int] = None,
+        filters: Optional[_Filters] = None,
+        group_by: Literal[None] = None,
+        generate: Literal[None] = None,
+        return_metadata: Optional[MetadataQuery] = None,
+        return_properties: Optional[Union[PROPERTIES, Type[Properties]]] = None,
+    ) -> _QueryReturn[Properties]:
+        ...
+
+    @overload
+    def near_text(
+        self,
+        query: Union[List[str], str],
+        certainty: Optional[float] = None,
+        distance: Optional[float] = None,
+        move_to: Optional[Move] = None,
+        move_away: Optional[Move] = None,
+        auto_limit: Optional[int] = None,
+        filters: Optional[_Filters] = None,
+        *,
+        group_by: Literal[None] = None,
+        generate: Generate,
+        return_metadata: Optional[MetadataQuery] = None,
+        return_properties: Optional[Union[PROPERTIES, Type[Properties]]] = None,
+    ) -> _GenerativeReturn[Properties]:
+        ...
+
+    @overload
+    def near_text(
+        self,
+        query: Union[List[str], str],
+        certainty: Optional[float] = None,
+        distance: Optional[float] = None,
+        move_to: Optional[Move] = None,
+        move_away: Optional[Move] = None,
+        auto_limit: Optional[int] = None,
+        filters: Optional[_Filters] = None,
+        *,
+        group_by: GroupBy,
+        generate: Literal[None] = None,
+        return_metadata: Optional[MetadataQuery] = None,
+        return_properties: Optional[Union[PROPERTIES, Type[Properties]]] = None,
+    ) -> _GroupByReturn[Properties]:
+        ...
+
+    def near_text(
+        self,
+        query: Union[List[str], str],
+        certainty: Optional[float] = None,
+        distance: Optional[float] = None,
+        move_to: Optional[Move] = None,
+        move_away: Optional[Move] = None,
+        auto_limit: Optional[int] = None,
+        filters: Optional[_Filters] = None,
+        group_by: Optional[GroupBy] = None,
+        generate: Optional[Generate] = None,
+        return_metadata: Optional[MetadataQuery] = None,
+        return_properties: Optional[Union[PROPERTIES, Type[Properties]]] = None,
+    ) -> Union[_QueryReturn[Properties], _GenerativeReturn[Properties], _GroupByReturn[Properties]]:
+        if generate is not None and group_by is not None:
+            raise ValueError("Cannot have group_by and generate defined simultaneously")
+
+        ret_properties, ret_type = self._determine_generic(return_properties)
+        res = self._query().near_text(
+            near_text=query,
+            certainty=certainty,
+            distance=distance,
+            move_to=move_to,
+            move_away=move_away,
+            autocut=auto_limit,
+            filters=filters,
+            group_by=group_by,
+            generate=generate,
+            return_metadata=return_metadata,
+            return_properties=ret_properties,
+        )
+        if generate is None and group_by is None:
+            return self._result_to_query_return(res, ret_type)
+        elif generate is not None:
+            return self._result_to_generative_return(res, ret_type)
+        else:
+            return self._result_to_groupby_return(res, ret_type)
